requirements.txt
aiogram>=3.0
openai-whisper
gTTS
